fix(tests): unblock the always-on corpus gates on mypy and file size
Two push gates failed on this branch. Both are fixed here without a
suppression and without raising a ratchet baseline.

mypy call-overload at test_always_on_corpus_claims.py:508.
`parse_corpus_prose` returns `dict[str, object]` because its values are
heterogeneous (ints plus a frozenset), so the annotation is honest and a
cast would be a lie. The bare `sorted(figures['members'])` had no way to
narrow. Added `assert isinstance(members, frozenset)`, which both
satisfies mypy and asserts a real parser contract the test depends on.

taste-lint file-size at 584 lines against a 500 line ceiling, which
pushed the repo taste count to 585 against a baseline of 584. The
linter's own remediation advice is to extract, so the constants,
regexes, and parsers moved to `tests/validation/always_on_corpus_helpers.py`,
following the existing `shipped_skill_routes_helpers.py` precedent. The
test module drops from 584 to 374 lines. Both files are under the
ceiling and the taste count is back at 584. The baseline is untouched.

Verification, all run after the split:
  ruff check                     All checks passed
  mypy (both files)              Success, no issues found
  taste_lints.py (both files)    0 errors
  taste_count_ratchet.py         OK (count == baseline 584)
  pytest (this module)           37 passed

Copilot-Session: 3ab110a4-2e9e-4504-a731-74fa73874896
